SecureGenericForum/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
SQLAlchemy==1.4.25
Werkzeug==2.0.2
uWSGI==2.0.20
six==1.16.0
pycparser==2.20
cffi==1.15.0
PyMySQL==1.0.2
mysql-connector-python==8.0.26
cryptography==35.0.0
qrcode==7.3.1
pyotp==2.6.0
pycryptodome==3.11.0
Pillow==8.4.0
behave==1.2.6
pytest==6.2.5
WTForms==2.3.3 
flask-wtf==0.15.1
flask_mailing==0.0.5
selenium==4.0.0
bleach==4.0.0
requests==2.26.0